requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8        
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
Flask==2.0.3
Flask-Cors==3.0.10
fonttools==4.31.2
get-all-tickers==1.8
idna==3.3
ipykernel==6.12.1
ipython==8.2.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.2.2
jupyter-core==4.9.2
kaleido==0.2.1
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
pandas-datareader==0.10.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
pymongo==4.0.2
pynpm==0.1.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
pywin32==303
pyxirr==0.7.2
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
stack-data==0.2.0
tenacity==8.0.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
XlsxWriter==3.0.3